add a sound in propulsion game played when win path is displayed
This sound is generated with a compressed sinusoid, and have no harmonics,
so I'm not sure at all if it may be pleasant (with some adjustement?)
For now, frequency used are from 220Hz to 440Hz (from A3 to A4), and may
be changed in editor if needed.
Need an additionnal check and maybe should be remove?